+++ src/backend/Domain/Entities/WeightEntry.cs
@@ -3,7 +3,7 @@ namespace Domain.Entities;
 public class WeightEntry : BaseEntity
 {
     // Value is in kilograms
-    public required double Value { get; set; }
+    public required decimal Value { get; set; }
     public string? Comment { get; set; }
     
     // yyyy-mm-dd format
